debt - move graph close to its only user
